requirement.txt
crewai
crewai[tools]
mdpdf
markdown
pdfkit
langchain-community
pyautogen[retrievechat]
poppler-utils
ipython